requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.1.1
youtube-transcript-api>=1.2.2
urllib3
